+++ tests/Assets/PhpFunctionOverrides.php
@@ -41,31 +41,3 @@ function header($string, $replace = true, $statusCode = null)
         ]
     );
 }
-
-/**
- * Is a file descriptor writable
- *
- * @param $file
- * @return bool
- */
-function is_readable($file)
-{
-    if (stripos($file, 'non-readable.cache') !== false) {
-        return false;
-    }
-    return true;
-}
-
-/**
- * Is a path writable
- *
- * @param $path
- * @return bool
- */
-function is_writable($path)
-{
-    if (stripos($path, 'non-writable-directory') !== false) {
-        return false;
-    }
-    return true;
-}
